Use math.ceil

+++ src/dls_bba/fbba.py
@@ -1,4 +1,5 @@
 import logging as log
+from math import ceil
 from typing import Any, Dict, List, Tuple
 
 import numpy as np
@@ -123,7 +124,9 @@ class FastBBA(Algorithm):
                 # This will block until all data has been retrieved.
                 fa_data = fa_buffer.get_data()
                 exc_data = (exc_high, exc_low)
-                selected_data = self.select_data(fa_data, components.axis, exc_data, decimated)
+                selected_data = self.select_data(
+                    fa_data, components.axis, exc_data, decimated
+                )
 
                 key = f"{quad_name}_{components.axis}_High"
                 rawdata[key] = selected_data[0]
@@ -180,7 +183,7 @@ class FastBBA(Algorithm):
         low_start = int(np.searchsorted(times, exc_low.start_time))
         log.debug("Searched start times: %s, %s", high_start, low_start)
         # Ensure we include the entire oscillation if using decimated data.
-        length = int(np.ceil(exc_high.count / 10)) if decimated else int(exc_high.count)
+        length = ceil(exc_high.count / 10) if decimated else exc_high.count
         high_data = data[high_start : high_start + length, :, plane]
         low_data = data[low_start : low_start + length, :, plane]
         log.debug("Selected data shape: {} {}".format(high_data.shape, low_data.shape))
